+++ source/graphics/2d/Textures/Char/CharSet.h
@@ -66,7 +66,7 @@
         /* char set definition */																		\
         CharSetDefinition* charSetDefinition;															\
         /* array definition of the charSet */															\
-        u16 charDefinitionDisplacement;																	\
+        u32 charDefinitionDisplacement;																	\
         /* memory displacement */																		\
         u16 offset;																						\
         /* how many textures are using me */															\
@@ -117,7 +117,7 @@ CharSetDefinition* CharSet_getCharSetDefinition(CharSet this);
 u32 CharSet_getNumberOfChars(CharSet this);
 void CharSet_write(CharSet this);
 void CharSet_rewrite(CharSet this);
-void CharSet_setCharDefinitionDisplacement(CharSet this, u16 charDefinitionDisplacement);
+void CharSet_setCharDefinitionDisplacement(CharSet this, u32 charDefinitionDisplacement);
 void CharSet_putChar(CharSet this, u32 charToReplace, BYTE* newChar);
 void CharSet_putPixel(CharSet this, u32 charToReplace, Point* charSetPixel, BYTE newPixelColor);
